requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.13.0
statsmodels>=0.13.5
openpyxl>=3.1.0
tabula-py>=2.7.0
scikit-learn>=1.0.2
# Uncomment the line below if you decide to integrate R directly via rpy2
# rpy2==3.4.5
